🔖 [dashboard 0.9.17] 404 页面

+++ model/common.go
@@ -14,7 +14,7 @@ type Common struct {
 }
 
 type Response struct {
-	Code    uint64      `json:"code,omitempty"`
+	Code    int         `json:"code,omitempty"`
 	Message string      `json:"message,omitempty"`
 	Result  interface{} `json:"result,omitempty"`
 }
